+++ extension_packages/figpack_franklab/src/index.tsx
@@ -2,6 +2,7 @@
 import React, { FunctionComponent, useEffect, useState } from "react";
 import { createRoot } from "react-dom/client";
 import {
+  DrawForExportFunction,
   FPViewComponent,
   FPViewContext,
   RenderParams,
@@ -91,6 +92,7 @@ type ComponentWrapperProps = {
   onDataChange: (callback: (zarrGroup: ZarrGroup) => void) => void;
   contexts: { [key: string]: FPViewContext };
   component: React.ComponentType<any>;
+  setDrawForExport?: (draw: DrawForExportFunction) => void;
 };
 
 // eslint-disable-next-line react-refresh/only-export-components
@@ -102,6 +104,7 @@ const ComponentWrapper: FunctionComponent<ComponentWrapperProps> = ({
   onDataChange,
   contexts,
   component: Component,
+  setDrawForExport,
 }) => {
   const [internalWidth, setInternalWidth] = useState(width);
   const [internalHeight, setInternalHeight] = useState(height);
@@ -123,6 +126,7 @@ const ComponentWrapper: FunctionComponent<ComponentWrapperProps> = ({
       width={internalWidth}
       height={internalHeight}
       contexts={contexts}
+      setDrawForExport={setDrawForExport}
     />
   );
 };
@@ -137,6 +141,7 @@ const makeRenderFunction = (Component: React.ComponentType<any>) => {
       onResize,
       onDataChange,
       contexts,
+      setDrawForExport,
     } = a;
     const root = createRoot(container);
     root.render(
@@ -148,6 +153,7 @@ const makeRenderFunction = (Component: React.ComponentType<any>) => {
         onDataChange={onDataChange}
         contexts={contexts}
         component={Component}
+        setDrawForExport={setDrawForExport}
       />,
     );
   };